# docs/server/swagger.json
{
    "components": {
        "schemas": {
...
            "pkg_api_v1.createSecretRequest": {
                "description": "Request to create a new secret",
                "properties": {
                    "key": {
                        "description": "Secret key name",
                        "type": "string"
                    },
                    "value": {
                        "description": "Secret value",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "pkg_api_v1.createSecretResponse": {
                "description": "Response after creating a secret",
                "properties": {
                    "key": {
                        "description": "Secret key that was created",
                        "type": "string"
                    },
                    "message": {
                        "description": "Success message",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "pkg_api_v1.getSecretsProviderResponse": {
                "description": "Response containing secrets provider details",
                "properties": {
                    "capabilities": {
                        "$ref": "#/components/schemas/pkg_api_v1.providerCapabilitiesResponse"
                    },
                    "name": {
                        "description": "Name of the secrets provider",
                        "type": "string"
                    },
                    "provider_type": {
                        "description": "Type of the secrets provider",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "pkg_api_v1.listSecretsResponse": {
                "description": "Response containing a list of secret keys",
                "properties": {
                    "keys": {
                        "description": "List of secret keys",
                        "items": {
                            "$ref": "#/components/schemas/pkg_api_v1.secretKeyResponse"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
...
            "pkg_api_v1.providerCapabilitiesResponse": {
                "description": "Capabilities of the secrets provider",
                "properties": {
                    "can_cleanup": {
                        "description": "Whether the provider can cleanup all secrets",
                        "type": "boolean"
                    },
                    "can_delete": {
                        "description": "Whether the provider can delete secrets",
                        "type": "boolean"
                    },
                    "can_list": {
                        "description": "Whether the provider can list secrets",
                        "type": "boolean"
                    },
                    "can_read": {
                        "description": "Whether the provider can read secrets",
                        "type": "boolean"
                    },
                    "can_write": {
                        "description": "Whether the provider can write secrets",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "pkg_api_v1.secretKeyResponse": {
                "description": "Secret key information",
                "properties": {
                    "description": {
                        "description": "Optional description of the secret",
                        "type": "string"
                    },
                    "key": {
                        "description": "Secret key name",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "pkg_api_v1.setupSecretsRequest": {
                "description": "Request to setup a secrets provider",
                "properties": {
                    "password": {
                        "description": "Password for encrypted provider (optional, can be set via environment variable)\nTODO Review environment variable for this",
                        "type": "string"
                    },
                    "provider_type": {
                        "description": "Type of the secrets provider (encrypted, 1password, environment)",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "pkg_api_v1.setupSecretsResponse": {
                "description": "Response after initializing a secrets provider",
                "properties": {
                    "message": {
                        "description": "Success message",
                        "type": "string"
                    },
                    "provider_type": {
                        "description": "Type of the secrets provider that was setup",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "info": {
        "description": "This is the ToolHive API server.",
        "title": "ToolHive API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/api/v1beta/secrets": {
            "post": {
                "description": "Setup the secrets provider with the specified type and configuration.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/pkg_api_v1.setupSecretsRequest",
                                        "summary": "request",
                                        "description": "Setup secrets provider request"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Setup secrets provider request",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/pkg_api_v1.setupSecretsResponse"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Setup or reconfigure secrets provider",
                "tags": [
                    "secrets"
                ]
            }
        },
        "/api/v1beta/secrets/default": {
            "get": {
                "description": "Get details of the default secrets provider",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/pkg_api_v1.getSecretsProviderResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Not Found - Provider not setup"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Get secrets provider details",
                "tags": [
                    "secrets"
                ]
            }
        },
        "/api/v1beta/secrets/default/keys": {
            "get": {
                "description": "Get a list of all secret keys from the default provider",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/pkg_api_v1.listSecretsResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Not Found - Provider not setup"
                    },
                    "405": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Method Not Allowed - Provider doesn't support listing"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "List secrets",
                "tags": [
                    "secrets"
                ]
            },
            "post": {
                "description": "Create a new secret in the default provider (encrypted provider only)",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/pkg_api_v1.createSecretRequest",
                                        "summary": "request",
                                        "description": "Create secret request"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Create secret request",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/pkg_api_v1.createSecretResponse"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Not Found - Provider not setup"
                    },
                    "405": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Method Not Allowed - Provider doesn't support writing"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Conflict - Secret already exists"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Create a new secret",
                "tags": [
                    "secrets"
                ]
            }
        },
...
    },
    "openapi": "3.1.0"
}